+++ src/locus.erl
@@ -36,6 +36,7 @@
 -export([get_version/1]).                 -ignore_xref({get_version,1}).
 -export([get_info/1]).                    -ignore_xref({get_info,1}).
 -export([get_info/2]).                    -ignore_xref({get_info,2}).
+-export([analyze_robustness/1]).          -ignore_xref({analyze_robustness,1}).
 
 -deprecated([{get_version,1,eventually}]).
 
@@ -307,6 +308,29 @@ get_info(DatabaseId, Property) ->
         {error, Error} ->
             {error, Error}
     end.
+
+%% @doc Analyzes a loaded database for corruption or incompatibility.
+%%
+%% <ul>
+%% <li>`DatabaseId' must be an atom and refer to a database loader.</li>
+%% </ul>
+%%
+%% Returns:
+%% <ul>
+%% <li>`ok' if the database is wholesome</li>
+%% <li>`{error, {frail, Frailties]}}' in case of corruption or incompatibility
+%%    (check the @link{locus_mmdb:frailty/0.definition of Frailty} for specifics.)
+%% </li>
+%% <li>`{error, database_unknown}' if the database loader for `DatabaseId' hasn't been started.</li>
+%% <li>`{error, database_not_loaded}' if the database hasn't yet been loaded.</li>
+%% </ul>
+-spec analyze_robustness(DatabaseId) -> ok | {error, Error}
+            when DatabaseId :: atom(),
+                 Error :: ({frail, [locus_mmdb:frailty(), ...]} |
+                           database_unknown |
+                           database_not_loaded).
+analyze_robustness(DatabaseId) ->
+    locus_mmdb:analyze_robustness(DatabaseId).
 
 %% ------------------------------------------------------------------
 %% CLI-only Function Definitions